// src/app/views/empresas/interfaces/empresa.ts
export interface Empresa {
    id: number;
    razon_social: string;
    nombre_comercial: string;
    registro_patronal: string;
    regimen_id:number,
    regimen: string;
    pais_id: number;
    pais: string;
    estado_id: number;
    estado: string;
    municipio_id: number;
    municipio: string;
    ciudad_id: number;
    ciudad: string;
    codigo_postal: string;
    calle: string;
    numero_exterior: string;
    numero_interior: string;

}